requirements.txt
alabaster==0.7.12
anyio==3.6.1
appdirs==1.4.4
appnope==0.1.3
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.0.8
async-generator==1.10
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bibtexparser==1.4.0
black==22.8.0
bleach==5.0.1
bump2version==0.5.11
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coverage==4.5.4
dataclass-wizard==0.22.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
docutils==0.17.1
entrypoints==0.3
exceptiongroup==1.0.0rc9
executing==1.0.0
fake-useragent==0.1.11
fastjsonschema==2.16.2
filelock==3.8.0
flake8==3.7.8
Flask==2.2.2
free-proxy==1.0.6
h11==0.14.0
html5lib==1.1
idna==3.4
imagesize==1.4.1
inflection==0.5.1
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.15.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==2.0.4
-e git+ssh://git@bitbucket.org/tbereau/my_scientific_profile.git@8c566a1c0b0e8825f39c827100a897608d5f62b4#egg=my_scientific_profile
mypy-extensions==0.4.3
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.3
orcid==1.0.3
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==0.13.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.5.0
pycparser==2.21
pydantic==1.10.2
pyflakes==2.1.1
Pygments==2.13.0
pyhumps==3.7.3
PyMuPDF==1.20.2
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.2.1
PyYAML==6.0
pyzmq==24.0.0
qtconsole==5.3.2
QtPy==2.2.0
readme-renderer==37.1
regex==2022.9.13
requests==2.28.1
requests-toolbelt==0.9.1
scholarly==1.7.2
selenium==4.5.0
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
Sphinx==1.8.5
sphinx-rtd-theme==1.0.0
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
stack-data==0.5.0
terminado==0.15.0
tinycss2==1.1.1
toml==0.10.2
tomli==1.2.3
tornado==6.2
tox==3.14.0
tqdm==4.64.1
traitlets==5.4.0
trio==0.22.0
trio-websocket==0.9.2
twine==1.14.0
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.5
watchdog==2.1.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
wsproto==1.2.0
